+++ src/Command - Copy/Dose.cs
@@ -1,6 +1,6 @@
-﻿/* PROJECT: Dose (https://github.com/aprettycoolprogram/Dose)
- *    FILE: Dose.Class1.cs
- * UPDATED: 6-30-2021-10:07 AM
+﻿/* PROJECT: Command (https://github.com/aprettycoolprogram/Command)
+ *    FILE: InptAdmitDate.Dose.cs
+ * UPDATED: 6-30-2021-10:18 AM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
@@ -11,9 +11,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Dose
+namespace Command
 {
-    public class Class1
+    class Dose
     {
     }
 }