+++ src/Command - Copy/TestFunctionality.cs
@@ -1,6 +1,6 @@
-﻿/* PROJECT: InptAdmitDate (https://github.com/aprettycoolprogram/InptAdmitDate)
- *    FILE: InptAdmitDate.Class1.cs
- * UPDATED: 6-30-2021-10:06 AM
+﻿/* PROJECT: Command (https://github.com/aprettycoolprogram/Command)
+ *    FILE: InptAdmitDate.TestFunctionality.cs
+ * UPDATED: 6-30-2021-11:55 AM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
@@ -11,9 +11,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace InptAdmitDate
+namespace Command
 {
-    public class Class1
+    public class TestFunctionality
     {
+        public static void ForceInptAdmitDate()
+        {
+
+        }
     }
 }